♻️ refactor: Card of Lending-Badge
